Update to Minecraft 26.1

+++ cerium/src/protocol/packet/server/play/block_update.rs
@@ -9,7 +9,7 @@ use crate::{
 #[derive(Debug, Clone)]
 pub struct BlockUpdatePacket {
     pub position: BlockPosition,
-    pub block_id: i32,
+    pub block_id: u16,
 }
 
 impl Packet for BlockUpdatePacket {}
@@ -18,7 +18,7 @@ impl ServerPacket for BlockUpdatePacket {}
 impl Encode for BlockUpdatePacket {
     fn encode<W: PacketWrite>(w: &mut W, this: &Self) -> Result<(), EncodeError> {
         w.write_position(&this.position)?;
-        w.write_varint(this.block_id)?;
+        w.write_varint(this.block_id as i32)?;
         Ok(())
     }
 }
